% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{Availability}
\alias{Availability}
\title{DoubleClick Search API Objects 
Reports and modifies your advertising data in DoubleClick Search (for example, campaigns, ad groups, keywords, and conversions).}
\usage{
Availability(advertiserId = NULL, agencyId = NULL,
  availabilityTimestamp = NULL, segmentationId = NULL,
  segmentationName = NULL, segmentationType = NULL)
}
\arguments{
\item{advertiserId}{DS advertiser ID}

\item{agencyId}{DS agency ID}

\item{availabilityTimestamp}{The time by which all conversions have been uploaded, in epoch millis UTC}

\item{segmentationId}{The numeric segmentation identifier (for example, DoubleClick Search Floodlight activity ID)}

\item{segmentationName}{The friendly segmentation identifier (for example, DoubleClick Search Floodlight activity name)}

\item{segmentationType}{The segmentation type that this availability is for (its default value is FLOODLIGHT)}
}
\value{
Availability object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:11:32
filename: /Users/mark/dev/R/autoGoogleAPI/googledoubleclicksearchv2.auto/R/doubleclicksearch_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Availability Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message containing availability data relevant to DoubleClick Search.
}

