% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_functions.R
\name{conversion.get}
\alias{conversion.get}
\title{Retrieves a list of conversions from a DoubleClick Search engine account.}
\usage{
conversion.get(agencyId, advertiserId, engineAccountId, endDate, rowCount,
  startDate, startRow, adGroupId = NULL, adId = NULL, campaignId = NULL,
  criterionId = NULL)
}
\arguments{
\item{agencyId}{Numeric ID of the agency}

\item{advertiserId}{Numeric ID of the advertiser}

\item{engineAccountId}{Numeric ID of the engine account}

\item{endDate}{Last date (inclusive) on which to retrieve conversions}

\item{rowCount}{The number of conversions to return per call}

\item{startDate}{First date (inclusive) on which to retrieve conversions}

\item{startRow}{The 0-based starting index for retrieving conversions results}

\item{adGroupId}{Numeric ID of the ad group}

\item{adId}{Numeric ID of the ad}

\item{campaignId}{Numeric ID of the campaign}

\item{criterionId}{Numeric ID of the criterion}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/doubleclicksearch
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/doubleclicksearch)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-search/}{Google Documentation}
}

