% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_functions.R
\name{conversion.patch}
\alias{conversion.patch}
\title{Updates a batch of conversions in DoubleClick Search. This method supports patch semantics.}
\usage{
conversion.patch(ConversionList, advertiserId, agencyId, endDate,
  engineAccountId, rowCount, startDate, startRow)
}
\arguments{
\item{ConversionList}{The \link{ConversionList} object to pass to this method}

\item{advertiserId}{Numeric ID of the advertiser}

\item{agencyId}{Numeric ID of the agency}

\item{endDate}{Last date (inclusive) on which to retrieve conversions}

\item{engineAccountId}{Numeric ID of the engine account}

\item{rowCount}{The number of conversions to return per call}

\item{startDate}{First date (inclusive) on which to retrieve conversions}

\item{startRow}{The 0-based starting index for retrieving conversions results}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/doubleclicksearch
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/doubleclicksearch)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-search/}{Google Documentation}

Other ConversionList functions: \code{\link{ConversionList}},
  \code{\link{conversion.insert}},
  \code{\link{conversion.update}}
}
