% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{Conversion}
\alias{Conversion}
\title{Conversion Object}
\usage{
Conversion(adGroupId = NULL, adId = NULL, advertiserId = NULL,
  agencyId = NULL, attributionModel = NULL, campaignId = NULL,
  channel = NULL, clickId = NULL, conversionId = NULL,
  conversionModifiedTimestamp = NULL, conversionTimestamp = NULL,
  countMillis = NULL, criterionId = NULL, currencyCode = NULL,
  customDimension = NULL, customMetric = NULL, deviceType = NULL,
  dsConversionId = NULL, engineAccountId = NULL, floodlightOrderId = NULL,
  inventoryAccountId = NULL, productCountry = NULL, productGroupId = NULL,
  productId = NULL, productLanguage = NULL, quantityMillis = NULL,
  revenueMicros = NULL, segmentationId = NULL, segmentationName = NULL,
  segmentationType = NULL, state = NULL, storeId = NULL, type = NULL)
}
\arguments{
\item{adGroupId}{DS ad group ID}

\item{adId}{DS ad ID}

\item{advertiserId}{DS advertiser ID}

\item{agencyId}{DS agency ID}

\item{attributionModel}{Available to advertisers only after contacting DoubleClick Search customer support}

\item{campaignId}{DS campaign ID}

\item{channel}{Sales channel for the product}

\item{clickId}{DS click ID for the conversion}

\item{conversionId}{For offline conversions, this is an ID that advertisers are required to provide}

\item{conversionModifiedTimestamp}{The time at which the conversion was last modified, in epoch millis UTC}

\item{conversionTimestamp}{The time at which the conversion took place, in epoch millis UTC}

\item{countMillis}{Available to advertisers only after contacting DoubleClick Search customer support}

\item{criterionId}{DS criterion (keyword) ID}

\item{currencyCode}{The currency code for the conversion's revenue}

\item{customDimension}{Custom dimensions for the conversion, which can be used to filter data in a report}

\item{customMetric}{Custom metrics for the conversion}

\item{deviceType}{The type of device on which the conversion occurred}

\item{dsConversionId}{ID that DoubleClick Search generates for each conversion}

\item{engineAccountId}{DS engine account ID}

\item{floodlightOrderId}{The Floodlight order ID provided by the advertiser for the conversion}

\item{inventoryAccountId}{ID that DS generates and uses to uniquely identify the inventory account that contains the product}

\item{productCountry}{The country registered for the Merchant Center feed that contains the product}

\item{productGroupId}{DS product group ID}

\item{productId}{The product ID (SKU)}

\item{productLanguage}{The language registered for the Merchant Center feed that contains the product}

\item{quantityMillis}{The quantity of this conversion, in millis}

\item{revenueMicros}{The revenue amount of this TRANSACTION conversion, in micros (value multiplied by 1000000, no decimal)}

\item{segmentationId}{The numeric segmentation identifier (for example, DoubleClick Search Floodlight activity ID)}

\item{segmentationName}{The friendly segmentation identifier (for example, DoubleClick Search Floodlight activity name)}

\item{segmentationType}{The segmentation type of this conversion (for example, FLOODLIGHT)}

\item{state}{The state of the conversion, that is, either ACTIVE or REMOVED}

\item{storeId}{The ID of the local store for which the product was advertised}

\item{type}{The type of the conversion, that is, either ACTION or TRANSACTION}
}
\value{
Conversion object
}
\description{
Conversion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A conversion containing data relevant to DoubleClick Search.
}

