% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_functions.R
\name{conversion.update}
\alias{conversion.update}
\title{Updates a batch of conversions in DoubleClick Search.}
\usage{
conversion.update(ConversionList)
}
\arguments{
\item{ConversionList}{The \link{ConversionList} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/doubleclicksearch
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/doubleclicksearch)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-search/}{Google Documentation}

Other ConversionList functions: \code{\link{ConversionList}},
  \code{\link{conversion.insert}},
  \code{\link{conversion.patch}}
}
