% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_functions.R
\name{conversion.updateAvailability}
\alias{conversion.updateAvailability}
\title{Updates the availabilities of a batch of floodlight activities in DoubleClick Search.}
\usage{
conversion.updateAvailability(UpdateAvailabilityRequest)
}
\arguments{
\item{UpdateAvailabilityRequest}{The \link{UpdateAvailabilityRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/doubleclicksearch
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/doubleclicksearch)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-search/}{Google Documentation}

Other UpdateAvailabilityRequest functions: \code{\link{UpdateAvailabilityRequest}}
}

