% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{ConversionList}
\alias{ConversionList}
\title{ConversionList Object}
\usage{
ConversionList(conversion = NULL)
}
\arguments{
\item{conversion}{The conversions being requested}
}
\value{
ConversionList object
}
\description{
ConversionList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of conversions.
}
\seealso{
Other ConversionList functions: \code{\link{conversion.insert}},
  \code{\link{conversion.patch}},
  \code{\link{conversion.update}}
}

