% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{CustomDimension}
\alias{CustomDimension}
\title{CustomDimension Object}
\usage{
CustomDimension(name = NULL, value = NULL)
}
\arguments{
\item{name}{Custom dimension name}

\item{value}{Custom dimension value}
}
\value{
CustomDimension object
}
\description{
CustomDimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message containing the custome dimension.
}

