% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{CustomMetric}
\alias{CustomMetric}
\title{CustomMetric Object}
\usage{
CustomMetric(name = NULL, value = NULL)
}
\arguments{
\item{name}{Custom metric name}

\item{value}{Custom metric numeric value}
}
\value{
CustomMetric object
}
\description{
CustomMetric Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message containing the custome metric.
}
