% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_functions.R
\docType{package}
\name{doubleclicksearch_googleAuthR}
\alias{doubleclicksearch_googleAuthR}
\alias{doubleclicksearch_googleAuthR-package}
\title{DoubleClick Search API
Reports and modifies your advertising data in DoubleClick Search (for example, campaigns, ad groups, keywords, and conversions).}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:11:32
filename: /Users/mark/dev/R/autoGoogleAPI/googledoubleclicksearchv2.auto/R/doubleclicksearch_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/doubleclicksearch
}
}

