% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{Report.files}
\alias{Report.files}
\title{Report.files Object}
\usage{
Report.files()
}
\value{
Report.files object
}
\description{
Report.files Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Report functions: \code{\link{Report}}
}

