% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{Report}
\alias{Report}
\title{Report Object}
\usage{
Report(Report.files = NULL, files = NULL, id = NULL,
  isReportReady = NULL, request = NULL, rowCount = NULL, rows = NULL,
  statisticsCurrencyCode = NULL, statisticsTimeZone = NULL)
}
\arguments{
\item{Report.files}{The \link{Report.files} object or list of objects}

\item{files}{Asynchronous report only}

\item{id}{Asynchronous report only}

\item{isReportReady}{Asynchronous report only}

\item{request}{The request that created the report}

\item{rowCount}{The number of report rows generated by the report, not including headers}

\item{rows}{Synchronous report only}

\item{statisticsCurrencyCode}{The currency code of all monetary values produced in the report, including values that are set by users (e}

\item{statisticsTimeZone}{If all statistics of the report are sourced from the same time zone, this would be it}
}
\value{
Report object
}
\description{
Report Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A DoubleClick Search report. This object contains the report request, some report metadata such as currency code, and the generated report rows or report files.
}
\seealso{
Other Report functions: \code{\link{Report.files}}
}

