% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{ReportApiColumnSpec}
\alias{ReportApiColumnSpec}
\title{ReportApiColumnSpec Object}
\usage{
ReportApiColumnSpec(columnName = NULL, customDimensionName = NULL,
  customMetricName = NULL, endDate = NULL, groupByColumn = NULL,
  headerText = NULL, platformSource = NULL,
  productReportPerspective = NULL, savedColumnName = NULL,
  startDate = NULL)
}
\arguments{
\item{columnName}{Name of a DoubleClick Search column to include in the report}

\item{customDimensionName}{Segments a report by a custom dimension}

\item{customMetricName}{Name of a custom metric to include in the report}

\item{endDate}{Inclusive day in YYYY-MM-DD format}

\item{groupByColumn}{Synchronous report only}

\item{headerText}{Text used to identify this column in the report output; defaults to columnName or savedColumnName when not specified}

\item{platformSource}{The platform that is used to provide data for the custom dimension}

\item{productReportPerspective}{Returns metrics only for a specific type of product activity}

\item{savedColumnName}{Name of a saved column to include in the report}

\item{startDate}{Inclusive date in YYYY-MM-DD format}
}
\value{
ReportApiColumnSpec object
}
\description{
ReportApiColumnSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request object used to create a DoubleClick Search report.
}

