% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{ReportRequest.filters}
\alias{ReportRequest.filters}
\title{ReportRequest.filters Object}
\usage{
ReportRequest.filters()
}
\value{
ReportRequest.filters object
}
\description{
ReportRequest.filters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ReportRequest functions: \code{\link{ReportRequest.orderBy}},
  \code{\link{ReportRequest.reportScope}},
  \code{\link{ReportRequest.timeRange}},
  \code{\link{ReportRequest}},
  \code{\link{reports.generate}},
  \code{\link{reports.request}}
}

