% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{ReportRequest.orderBy}
\alias{ReportRequest.orderBy}
\title{ReportRequest.orderBy Object}
\usage{
ReportRequest.orderBy()
}
\value{
ReportRequest.orderBy object
}
\description{
ReportRequest.orderBy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ReportRequest functions: \code{\link{ReportRequest.filters}},
  \code{\link{ReportRequest.reportScope}},
  \code{\link{ReportRequest.timeRange}},
  \code{\link{ReportRequest}},
  \code{\link{reports.generate}},
  \code{\link{reports.request}}
}
