% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{ReportRequest}
\alias{ReportRequest}
\title{ReportRequest Object}
\usage{
ReportRequest(ReportRequest.filters = NULL, ReportRequest.orderBy = NULL,
  ReportRequest.reportScope = NULL, ReportRequest.timeRange = NULL,
  columns = NULL, downloadFormat = NULL, filters = NULL,
  includeDeletedEntities = NULL, includeRemovedEntities = NULL,
  maxRowsPerFile = NULL, orderBy = NULL, reportScope = NULL,
  reportType = NULL, rowCount = NULL, startRow = NULL,
  statisticsCurrency = NULL, timeRange = NULL,
  verifySingleTimeZone = NULL)
}
\arguments{
\item{ReportRequest.filters}{The \link{ReportRequest.filters} object or list of objects}

\item{ReportRequest.orderBy}{The \link{ReportRequest.orderBy} object or list of objects}

\item{ReportRequest.reportScope}{The \link{ReportRequest.reportScope} object or list of objects}

\item{ReportRequest.timeRange}{The \link{ReportRequest.timeRange} object or list of objects}

\item{columns}{The columns to include in the report}

\item{downloadFormat}{Format that the report should be returned in}

\item{filters}{A list of filters to be applied to the report}

\item{includeDeletedEntities}{Determines if removed entities should be included in the report}

\item{includeRemovedEntities}{Determines if removed entities should be included in the report}

\item{maxRowsPerFile}{Asynchronous report only}

\item{orderBy}{Synchronous report only}

\item{reportScope}{The reportScope is a set of IDs that are used to determine which subset of entities will be returned in the report}

\item{reportType}{Determines the type of rows that are returned in the report}

\item{rowCount}{Synchronous report only}

\item{startRow}{Synchronous report only}

\item{statisticsCurrency}{Specifies the currency in which monetary will be returned}

\item{timeRange}{If metrics are requested in a report, this argument will be used to restrict the metrics to a specific time range}

\item{verifySingleTimeZone}{If true, the report would only be created if all the requested stat data are sourced from a single timezone}
}
\value{
ReportRequest object
}
\description{
ReportRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A request object used to create a DoubleClick Search report.
}
\seealso{
Other ReportRequest functions: \code{\link{ReportRequest.filters}},
  \code{\link{ReportRequest.orderBy}},
  \code{\link{ReportRequest.reportScope}},
  \code{\link{ReportRequest.timeRange}},
  \code{\link{reports.generate}},
  \code{\link{reports.request}}
}
