% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{ReportRequest.reportScope}
\alias{ReportRequest.reportScope}
\title{ReportRequest.reportScope Object}
\usage{
ReportRequest.reportScope(adGroupId = NULL, adId = NULL,
  advertiserId = NULL, agencyId = NULL, campaignId = NULL,
  engineAccountId = NULL, keywordId = NULL)
}
\arguments{
\item{adGroupId}{DS ad group ID}

\item{adId}{DS ad ID}

\item{advertiserId}{DS advertiser ID}

\item{agencyId}{DS agency ID}

\item{campaignId}{DS campaign ID}

\item{engineAccountId}{DS engine account ID}

\item{keywordId}{DS keyword ID}
}
\value{
ReportRequest.reportScope object
}
\description{
ReportRequest.reportScope Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The reportScope is a set of IDs that are used to determine which subset of entities will be returned in the report. The full lineage of IDs from the lowest scoped level desired up through agency is required.
}
\seealso{
Other ReportRequest functions: \code{\link{ReportRequest.filters}},
  \code{\link{ReportRequest.orderBy}},
  \code{\link{ReportRequest.timeRange}},
  \code{\link{ReportRequest}},
  \code{\link{reports.generate}},
  \code{\link{reports.request}}
}
