% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{ReportRequest.timeRange}
\alias{ReportRequest.timeRange}
\title{ReportRequest.timeRange Object}
\usage{
ReportRequest.timeRange(changedAttributesSinceTimestamp = NULL,
  changedMetricsSinceTimestamp = NULL, endDate = NULL, startDate = NULL)
}
\arguments{
\item{changedAttributesSinceTimestamp}{Inclusive UTC timestamp in RFC format, e}

\item{changedMetricsSinceTimestamp}{Inclusive UTC timestamp in RFC format, e}

\item{endDate}{Inclusive date in YYYY-MM-DD format}

\item{startDate}{Inclusive date in YYYY-MM-DD format}
}
\value{
ReportRequest.timeRange object
}
\description{
ReportRequest.timeRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
If metrics are requested in a report, this argument will be used to restrict the metrics to a specific time range.
}
\seealso{
Other ReportRequest functions: \code{\link{ReportRequest.filters}},
  \code{\link{ReportRequest.orderBy}},
  \code{\link{ReportRequest.reportScope}},
  \code{\link{ReportRequest}},
  \code{\link{reports.generate}},
  \code{\link{reports.request}}
}

