% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{ReportRow}
\alias{ReportRow}
\title{ReportRow Object}
\usage{
ReportRow()
}
\value{
ReportRow object
}
\description{
ReportRow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A row in a DoubleClick Search report.
}
