% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_functions.R
\name{reports.generate}
\alias{reports.generate}
\title{Generates and returns a report immediately.}
\usage{
reports.generate(ReportRequest)
}
\arguments{
\item{ReportRequest}{The \link{ReportRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/doubleclicksearch
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/doubleclicksearch)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-search/}{Google Documentation}

Other ReportRequest functions: \code{\link{ReportRequest.filters}},
  \code{\link{ReportRequest.orderBy}},
  \code{\link{ReportRequest.reportScope}},
  \code{\link{ReportRequest.timeRange}},
  \code{\link{ReportRequest}},
  \code{\link{reports.request}}
}
