% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_functions.R
\name{reports.get}
\alias{reports.get}
\title{Polls for the status of a report request.}
\usage{
reports.get(reportId)
}
\arguments{
\item{reportId}{ID of the report request being polled}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/doubleclicksearch
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/doubleclicksearch)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-search/}{Google Documentation}
}
