% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{SavedColumn}
\alias{SavedColumn}
\title{SavedColumn Object}
\usage{
SavedColumn(savedColumnName = NULL, type = NULL)
}
\arguments{
\item{savedColumnName}{The name of the saved column}

\item{type}{The type of data this saved column will produce}
}
\value{
SavedColumn object
}
\description{
SavedColumn Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A saved column
}

