% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{SavedColumnList}
\alias{SavedColumnList}
\title{SavedColumnList Object}
\usage{
SavedColumnList(items = NULL)
}
\arguments{
\item{items}{The saved columns being requested}
}
\value{
SavedColumnList object
}
\description{
SavedColumnList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of saved columns. Advertisers create saved columns to report on Floodlight activities, Google Analytics goals, or custom KPIs. To request reports with saved columns, you'll need the saved column names that are available from this list.
}
