% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_functions.R
\name{savedColumns.list}
\alias{savedColumns.list}
\title{Retrieve the list of saved columns for a specified advertiser.}
\usage{
savedColumns.list(agencyId, advertiserId)
}
\arguments{
\item{agencyId}{DS ID of the agency}

\item{advertiserId}{DS ID of the advertiser}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/doubleclicksearch
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/doubleclicksearch)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-search/}{Google Documentation}
}

