% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{UpdateAvailabilityRequest}
\alias{UpdateAvailabilityRequest}
\title{UpdateAvailabilityRequest Object}
\usage{
UpdateAvailabilityRequest(availabilities = NULL)
}
\arguments{
\item{availabilities}{The availabilities being requested}
}
\value{
UpdateAvailabilityRequest object
}
\description{
UpdateAvailabilityRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request to update availability.
}
\seealso{
Other UpdateAvailabilityRequest functions: \code{\link{conversion.updateAvailability}}
}

