% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclicksearch_objects.R
\name{UpdateAvailabilityResponse}
\alias{UpdateAvailabilityResponse}
\title{UpdateAvailabilityResponse Object}
\usage{
UpdateAvailabilityResponse(availabilities = NULL)
}
\arguments{
\item{availabilities}{The availabilities being returned}
}
\value{
UpdateAvailabilityResponse object
}
\description{
UpdateAvailabilityResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a update availability request.
}
